+++ integration/ibc/cmd_ibc_test.go
@@ -189,7 +189,7 @@ func TestCreateIBCOracle(t *testing.T) {
 func TestCreateIBCPacket(t *testing.T) {
 	var (
 		env = envtest.New(t)
-		app = env.Scaffold("github.com/test/blogibc2")
+		app = env.Scaffold("github.com/test/blogibcb")
 	)
 
 	env.Must(env.Exec("create an IBC module",